+++ amethyst_ui/src/prefab.rs
@@ -6,10 +6,7 @@ use amethyst_audio::{AudioFormat, Source as Audio};
 use amethyst_core::specs::prelude::{Entities, Entity, Read, ReadExpect, Write, WriteStorage};
 use amethyst_renderer::{Texture, TextureFormat, TextureMetadata, TexturePrefab};
 use serde::de::DeserializeOwned;
-use {
-    Anchor, FontAsset, FontFormat, MouseReactive, OnUiActionImage, OnUiActionSound, Stretch,
-    TextEditing, UiButton, UiFocused, UiImage, UiText, UiTransform,
-};
+use super::*;
 
 /// Loadable `UiTransform` data.
 /// By default z is equal to one.
@@ -164,9 +161,13 @@ where
     pub color: [f32; 4],
     /// Font
     pub font: AssetPrefab<FontAsset, F>,
-    /// Password field ?
+    /// Should the text be shown as dots instead of the proper characters?
     #[serde(default)]
     pub password: bool,
+    /// Where should the text be aligned from. Relative to its own UiTransform's area.
+    pub align: Option<Anchor>,
+    /// How should the text behave with line breaks.
+    pub line_mode: Option<LineMode>,
     /// Optionally make the text editable
     #[serde(default)]
     pub editable: Option<TextEditingPrefab>,
@@ -222,6 +223,15 @@ where
         let font_handle = self.font.load_prefab(entity, fonts, &[])?;
         let mut ui_text = UiText::new(font_handle, self.text.clone(), self.color, self.font_size);
         ui_text.password = self.password;
+
+        if let Some(ref align) = self.align {
+             ui_text.align = align.clone();
+         }
+         
+         if let Some(ref line_mode) = self.line_mode {
+             ui_text.line_mode = line_mode.clone();
+         }
+
         texts.insert(entity, ui_text)?;
         if let Some(ref editing) = self.editable {
             editables.insert(
@@ -368,7 +378,9 @@ where
             ref mut textures,
             ref mut sounds,
         ) = system_data;
-        let normal_image = self.normal_image.load_prefab(entity, textures, entity_set)?;
+        let normal_image = self
+            .normal_image
+            .load_prefab(entity, textures, entity_set)?;
         let hover_image = self.hover_image.load_prefab(entity, textures, entity_set)?;
         let press_image = self.press_image.load_prefab(entity, textures, entity_set)?;
         let hover_sound = self.hover_sound.load_prefab(entity, sounds, entity_set)?;
@@ -501,7 +513,8 @@ where
     fn import(&self, bytes: Vec<u8>, _: ()) -> AssetResult<UiPrefab<A, I, F>> {
         use ron::de::Deserializer;
         use serde::Deserialize;
-        let mut d = Deserializer::from_bytes(&bytes).chain_err(|| "Failed deserializing Ron file")?;
+        let mut d =
+            Deserializer::from_bytes(&bytes).chain_err(|| "Failed deserializing Ron file")?;
         let root: UiWidget<A, I, F> =
             UiWidget::deserialize(&mut d).chain_err(|| "Failed parsing Ron file")?;
         d.end().chain_err(|| "Failed parsing Ron file")?;
@@ -565,6 +578,8 @@ fn walk_ui_tree<A, I, F>(
                 editable: None,
                 font: button.font.clone(),
                 password: false,
+                align: None,
+                line_mode: None,
                 text: button.text.clone(),
                 font_size: button.font_size,
             };
@@ -706,6 +721,5 @@ fn button_text_transform(mut id: String) -> UiTransformBuilder {
         .with_stretch(Stretch::XY {
             x_margin: 0.,
             y_margin: 0.,
-        })
-        .transparent()
+        }).transparent()
 }